KYI_Module/requirements.txt
langchain
PyPDF
langchain_openai
langgraph 
matplotlib
langchain_community
openai
python-dotenv
fastapi
supabase
surya-ocr
supabase
pyjwt
transformers
resend
supabase
uvicorn